refactor: Hand -> Cards, Participant -> Hand,
