+++ src/main/kotlin/blackjack/domain/Hand.kt
@@ -1,23 +1,23 @@
 package blackjack.domain
 
-class Hand {
-    private val _value: MutableList<Card> = mutableListOf()
-    val value: List<Card> get() = _value
+class Hand(
+    private val onBusted: () -> Unit,
+) {
+    private val _cards = Cards()
+    val cards: List<Card> = _cards.value
 
-    /**
-     * @return null if score is bigger than 21
-     * */
-    val score: Int?
-        get() {
-            if (value.isEmpty()) return 0
-            val possibleScores = ScoreCalculator.possibleScoreOf(*(value.toTypedArray()))
-            return possibleScores.sortedDescending().firstOrNull { score: Int -> score <= 21 }
-        }
+    val score: Int
+        get() =
+            _cards.score ?: run {
+                onBusted()
+                SCORE_BUSTED
+            }
 
-    fun add(card: Card) {
-        require(canGetCard()) { "모든 카드의 합이 21 미만이 될 수 있을 경우에만 카드를 얻을 수 있습니다." }
-        _value.add(card)
+    fun draw(card: Card) {
+        _cards.add(card)
     }
 
-    private fun canGetCard(): Boolean = score != null && score != 21
+    companion object {
+        private const val SCORE_BUSTED = -1
+    }
 }
